GUI and algorithm implementation
ACTUALLY finished the GUI this time. At least the bulk of it. There's a weird thing that happens when the question labels and textboxes pop up that I ask the TA about in lab tomorrow.

Algorithm (or whatever you call it... function) has been started. Jaz will implement the rest